# More scalable rwlock implementation