This patch implements more scalable version of the reader-writer lock. It should specifically improve performance of read-heavy workflows.

The implementation is based on similar ideas used in the reader-writer lock implementation in Golang and described in this article by Eli Bendersky - "A more efficient writer-preferring RW lock" in https://eli.thegreenplace.net/2019/implementing-reader-writer-locks/. However, it uses OSv primitives - wait()/wake() - to facilitate synchronization between writer and readers. Also, I believe, it uses a novel idea of using single field - `_readers` - to count both owning readers and pending readers at the same time (it could be that I re-invented what somebody else already invented before me). Last but not least, I was inspired by another implementation using atomics only and thread::yield() by Christian Dietrich (see https://gitlab.ibr.cs.tu-bs.de/kanoun/osv/-/blob/dbdb91c1c7d1a16c7ad2093c0f61fe5914656f2d/core/rwlock.cc)

In essence, this implementation uses a mutex to synchronize access between writers and atomics to synchronize access between readers and writers.

This RW lock should be fair for both writers and readers. Even though the pending writer sets WRITE_INDICATOR bit to block new readers, the owning writer in wunlock() wakes pending readers to acquire the lock for reading. Therefore, neither readers nor writers should starve.

Here are some tests results that compare various metrics of this new implementation with the original one:

```
=== misc-mutex ===
=== Uncontended single-thread lock/unlock cycle
=================
Original
--------
Measuring uncontended lockfree::mutex lock/unlock: 16 ns
Measuring uncontended handoff_stressing_mutex lock/unlock: 34 ns
Measuring uncontended spinlock lock/unlock: 8 ns
Measuring uncontended rwlock_read_lock lock/unlock: 36 ns
Measuring uncontended rwlock_write_lock lock/unlock: 39 ns ----
New
----
Measuring uncontended lockfree::mutex lock/unlock: 16 ns
Measuring uncontended handoff_stressing_mutex lock/unlock: 36 ns
Measuring uncontended spinlock lock/unlock: 8 ns
Measuring uncontended rwlock_read_lock lock/unlock: 16 ns
Measuring uncontended rwlock_write_lock lock/unlock: 40 ns
```

The reader path or rwlock in the uncontended cases is at least twice cheaper.

```
=== Contended lock/unlock scenarios
========
Original
--------
Contended mutex test, rwlock_write_lock, 2 pinned threads ABBABABABABABABABABA
1712 ns
Contended mutex test, rwlock_write_lock, 4 pinned threads ADBCADBCADCBADCBADCBADCBABDCADBCADBCABDC
1988 ns
Contended mutex test, rwlock_write_lock, 20 non-pinned threads 1272 ns
Contended mutex test, rwlock_read_lock, 2 pinned threads ABBABAABBABABAABBABA
1193 ns
Contended mutex test, rwlock_read_lock, 4 pinned threads ABCDABCDABCDABCDABCDABCDABCDABCDABCDABCD
1229 ns
Contended mutex test, rwlock_read_lock, 20 non-pinned threads 895 ns
---
New
---
Contended mutex test, rwlock_write_lock, 2 pinned threads ABBAABBABAABBABAABBA
603 ns
Contended mutex test, rwlock_write_lock, 4 pinned threads ABDCABCDABCDABCDABCDABCDABCDACBDABCDABCD
609 ns
Contended mutex test, rwlock_write_lock, 20 non-pinned threads 470 ns
Contended mutex test, rwlock_read_lock, 2 pinned threads ABBAABABABABABABABAB
86 ns
Contended mutex test, rwlock_read_lock, 4 pinned threads ABCDDDDDDDDDDBCACABCABCABCACBACBACACBABB
112 ns
Contended mutex test, rwlock_read_lock, 20 non-pinned threads 46 ns
```
As one can see the reader path in the contended case is dramatically faster.

```
Other misc-mutex test
--
Original
---
Mutual exclusion test using checker_thread with rwlock_write_lock:
Contended mutex test, rwlock_write_lock, 2 pinned threads 1547 ns

Contended mutex test, rwlock_write_lock, 4 pinned threads 1625 ns

Contended mutex test, rwlock_write_lock, 20 non-pinned threads 2092 ns

--
New
---
Mutual exclusion test using checker_thread with rwlock_write_lock:
Contended mutex test, rwlock_write_lock, 2 pinned threads 1223 ns

Contended mutex test, rwlock_write_lock, 4 pinned threads 1274 ns

Contended mutex test, rwlock_write_lock, 20 non-pinned threads 1138 ns
```
```
====
New reader-writer test with both readers and writers:
-- Original
Cmdline: /tests/tst-rwlock.so
Running contended read-write lock tests

Test with 10 writers and 1000 readers, 100 iterations, 1 writer inner loops, data len=1000, non-pinned threads
Reader lock/unlock count = 100000, lock wait: total = 1121640.17 ms, avg = 11216401.68 ns, unlock wait: total = 131800.30 ms, avg = 1318003.05 ns
Writer lock/unlock count = 1000, lock wait: total = 7136.79 ms, avg = 7136788.35 ns, unlock wait: total = 1159.55 ms, avg = 1159546.43 ns Total runtime 1603.18 ms

Test with 10 writers and 1000 readers, 100 iterations, 1 writer inner loops, data len=1000, pinned threads
Reader lock/unlock count = 100000, lock wait: total = 2197968.93 ms, avg = 21979689.31 ns, unlock wait: total = 158759.70 ms, avg = 1587597.02 ns
Writer lock/unlock count = 1000, lock wait: total = 16269.13 ms, avg = 16269125.97 ns, unlock wait: total = 2007.15 ms, avg = 2007152.40 ns Total runtime 2407.85 ms

-- New
Cmdline: /tests/tst-rwlock.so
Running contended read-write lock tests

Test with 10 writers and 1000 readers, 100 iterations, 1 writer inner loops, data len=1000, non-pinned threads
Reader lock/unlock count = 100000, lock wait: total = 137.68 ms, avg = 1376.83 ns, unlock wait: total = 8.65 ms, avg = 86.47 ns
Writer lock/unlock count = 1000, lock wait: total = 4.58 ms, avg = 4582.71 ns, unlock wait: total = 0.32 ms, avg = 319.08 ns Total runtime 524.06 ms

Test with 10 writers and 1000 readers, 100 iterations, 1 writer inner loops, data len=1000, pinned threads
Reader lock/unlock count = 100000, lock wait: total = 17.04 ms, avg = 170.40 ns, unlock wait: total = 7.52 ms, avg = 75.19 ns
Writer lock/unlock count = 1000, lock wait: total = 2.17 ms, avg = 2173.15 ns, unlock wait: total = 0.75 ms, avg = 745.52 ns Total runtime 390.06 ms
```
MySQL tests:
```
==
Original
----
Threads started!

[ 10s ] thds: 12 tps: 382.03 qps: 7653.88 (r/w/o: 5359.68/1528.94/765.27) lat (ms,95%): 44.98 err/s: 0.00 reconn/s: 0.00 [ 20s ] thds: 12 tps: 404.20 qps: 8083.40 (r/w/o: 5658.47/1616.52/808.41) lat (ms,95%): 40.37 err/s: 0.00 reconn/s: 0.00 [ 30s ] thds: 12 tps: 407.50 qps: 8153.40 (r/w/o: 5708.30/1630.10/815.00) lat (ms,95%): 40.37 err/s: 0.00 reconn/s: 0.00 [ 40s ] thds: 12 tps: 392.80 qps: 7852.54 (r/w/o: 5495.76/1571.19/785.59) lat (ms,95%): 41.85 err/s: 0.00 reconn/s: 0.00 [ 50s ] thds: 12 tps: 360.50 qps: 7210.39 (r/w/o: 5047.49/1441.90/721.00) lat (ms,95%): 44.17 err/s: 0.00 reconn/s: 0.00 [ 60s ] thds: 12 tps: 364.70 qps: 7292.50 (r/w/o: 5104.77/1458.32/729.41) lat (ms,95%): 44.98 err/s: 0.00 reconn/s: 0.00 SQL statistics:
    queries performed:
        read:                            323820
        write:                           92520
        other:                           46260
        total:                           462600
    transactions:                        23130  (385.37 per sec.)
    queries:                             462600 (7707.31 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          60.0195s
    total number of events:              23130

Latency (ms):
         min:                                    5.75
         avg:                                   31.13
         max:                                  116.35
         95th percentile:                       42.61
         sum:                               720082.28

Threads fairness:
    events (avg/stddev):           1927.5000/17.10
    execution time (avg/stddev):   60.0069/0.00

New
----
Threads started!

[ 10s ] thds: 12 tps: 427.02 qps: 8553.25 (r/w/o: 5989.65/1708.37/855.24) lat (ms,95%): 41.10 err/s: 0.00 reconn/s: 0.00 [ 20s ] thds: 12 tps: 417.81 qps: 8354.94 (r/w/o: 5847.77/1671.65/835.52) lat (ms,95%): 38.94 err/s: 0.00 reconn/s: 0.00 [ 30s ] thds: 12 tps: 408.70 qps: 8171.71 (r/w/o: 5719.84/1634.48/817.39) lat (ms,95%): 39.65 err/s: 0.00 reconn/s: 0.00 [ 40s ] thds: 12 tps: 412.99 qps: 8261.22 (r/w/o: 5782.87/1652.26/826.08) lat (ms,95%): 38.94 err/s: 0.00 reconn/s: 0.00 [ 50s ] thds: 12 tps: 413.81 qps: 8282.45 (r/w/o: 5799.51/1655.33/827.62) lat (ms,95%): 40.37 err/s: 0.00 reconn/s: 0.00 [ 60s ] thds: 12 tps: 415.70 qps: 8313.94 (r/w/o: 5819.73/1662.81/831.40) lat (ms,95%): 39.65 err/s: 0.00 reconn/s: 0.00 SQL statistics:
    queries performed:
        read:                            349622
        write:                           99892
        other:                           49946
        total:                           499460
    transactions:                        24973  (416.10 per sec.)
    queries:                             499460 (8321.99 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          60.0155s
    total number of events:              24973

Latency (ms):
         min:                                    5.34
         avg:                                   28.83
         max:                                  106.06
         95th percentile:                       39.65
         sum:                               720051.48

Threads fairness:
    events (avg/stddev):           2081.0833/22.89
    execution time (avg/stddev):   60.0043/0.00
```

The MySQL test using new rwlock is 8-10 % faster.